requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.10.0